dtos,mapeacao, validacao de campo

+++ ProEventos.API/Controllers/EventosController.cs
@@ -4,6 +4,7 @@ using System;
 using System.Threading.Tasks;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.DTOs;
 
 namespace ProEventos.API.Controllers
 {
@@ -25,7 +26,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -42,7 +43,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetEventoByIdAsync(id, true);
-                if (eventos == null) return NotFound("Eventos por id não encontrados");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -58,7 +59,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (eventos == null) return NotFound("Eventos por tema não encontrados");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -69,12 +70,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var eventos = await _eventoService.AddEventos(model);
-                if (eventos == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -85,12 +86,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var eventos = await _eventoService.UpdateEvento(id, model);
-                if (eventos == null) return BadRequest("Erro ao tentar atualizar evento.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -105,9 +106,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var eventos = await _eventoService.GetEventoByIdAsync(id, true);
+                if (eventos == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                     Ok("Deletado") : 
-                    BadRequest("Evento nãod deletado");
+                    throw new Exception ("Ocorreu um probelma não específico ao tentar deletar o Evento");
             }
             catch (Exception ex)
             {
